test(sessions): satisfy backend size limits
