+++ tests/functional/sessions/controls/helpers_test.go
@@ -14,7 +14,6 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	modelprovider "github.com/portpowered/infinite-you/pkg/services/models"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"github.com/portpowered/infinite-you/tests/functional/internal/support"
 )
@@ -898,105 +897,6 @@ func assertPauseResumeLifecycleControlEvents(
 			resumePayload.OccurredAt.UTC(),
 		)
 	}
-}
-
-func invocationPipelineConfig() map[string]any {
-	return map[string]any{
-		"workTypes": []map[string]any{{
-			"name": "task",
-			"states": []map[string]string{
-				{"name": "init", "type": "INITIAL"},
-				{"name": "complete", "type": "TERMINAL"},
-				{"name": "failed", "type": "FAILED"},
-			},
-		}},
-		"workers": []map[string]string{{"name": "worker-a"}},
-		"workstations": []map[string]any{{
-			"name":      "process",
-			"worker":    "worker-a",
-			"inputs":    []map[string]string{{"workType": "task", "state": "init"}},
-			"outputs":   []map[string]string{{"workType": "task", "state": "complete"}},
-			"onFailure": []map[string]string{{"workType": "task", "state": "failed"}},
-		}},
-	}
-}
-
-func scaffoldInvocationFactory(t *testing.T, overrides map[string]any) string {
-	t.Helper()
-
-	cfg := invocationPipelineConfig()
-	for key, value := range overrides {
-		cfg[key] = value
-	}
-	workTypes := cfg["workTypes"].([]map[string]any)
-	for i := range workTypes {
-		if name, _ := workTypes[i]["name"].(string); name == "task" {
-			workTypes[i]["handlingBehavior"] = []string{"DEFAULT"}
-		}
-	}
-	dir := support.ScaffoldFactory(t, cfg)
-	support.WriteAgentConfig(t, dir, "worker-a", support.BuildModelWorkerConfig(modelprovider.ProviderCodex, "gpt-5-codex"))
-	return dir
-}
-
-func textInvocationRequest(t *testing.T, text string, timeoutMillis *int64) factoryapi.InvocationRequest {
-	t.Helper()
-
-	var part factoryapi.WorkContentPart
-	if err := part.FromWorkTextContentPart(factoryapi.WorkTextContentPart{
-		Type: factoryapi.WorkContentPartTypeText,
-		Text: text,
-	}); err != nil {
-		t.Fatalf("build invocation text content: %v", err)
-	}
-	sourceKind := factoryapi.InvocationInputSourceKindText
-	content := factoryapi.WorkContent{part}
-	return factoryapi.InvocationRequest{
-		SourceKind:    &sourceKind,
-		Content:       &content,
-		TimeoutMillis: timeoutMillis,
-	}
-}
-
-func postInvocation(
-	t *testing.T,
-	serverURL string,
-	sessionID string,
-	request factoryapi.InvocationRequest,
-) factoryapi.InvocationResponse {
-	t.Helper()
-
-	body, err := json.Marshal(request)
-	if err != nil {
-		t.Fatalf("marshal invocation request: %v", err)
-	}
-	response, err := http.Post(
-		controlsSessionEndpoint(serverURL, sessionID, "/invocations"),
-		"application/json",
-		bytes.NewReader(body),
-	)
-	if err != nil {
-		t.Fatalf("POST /factory-sessions/%s/invocations: %v", sessionID, err)
-	}
-	defer response.Body.Close()
-	if response.StatusCode != http.StatusOK {
-		payload, readErr := io.ReadAll(response.Body)
-		if readErr != nil {
-			t.Fatalf("read invocation error response: %v", readErr)
-		}
-		t.Fatalf(
-			"POST /factory-sessions/%s/invocations status = %d: %s",
-			sessionID,
-			response.StatusCode,
-			strings.TrimSpace(string(payload)),
-		)
-	}
-
-	var decoded factoryapi.InvocationResponse
-	if err := json.NewDecoder(response.Body).Decode(&decoded); err != nil {
-		t.Fatalf("decode invocation response: %v", err)
-	}
-	return decoded
 }
 
 func scaffoldInterruptedInspectFactory(t *testing.T) string {
